Refactor database repository structure and implement base repository class
- Moved repository imports to a dedicated 'repositories' directory for better organization.
